Added insights to newly added analysis, updated insights fo existing ones. There few more insights to add and few more analysis to be done.
removed zip code analysis as it was done by grouping zip codes. This cannot be done since zip codes are basically address and address cannot be grouped in this way for analysis
removed open account and total accounts analysis. There are cases that shows 100 percent defaults since there was only one loan in that category and that is defaulted. For a better analysis this should be grouped into buckets